If in doubt, transfer a file in a `...load_if_newer` call.
This implementation considers the precisions of the remote and
local files. For example, if the remote and local file formally
have the same modification datetime, the remote file may still
be newer because its modification datateime can be within a
minute interval from the timestamp to the timestamp plus the
precision.